Changed Carrier Label to blue
Update:  A small amount of cleanup, and reduced the size
         of the carrier label to be less intrusive

PD and I were talking, and think it might be interesting
to have it show RED when the phone is ROAMING, and blue
normally.

I will push that in a future commit when I figure out
how Im going to accomplish that.

Change-Id: Iadecb4f92dd6eefedfa1681fe86747652986468b