Fix date themeability.
We should hunt down any hardcoded color strings that should be set
with a color resource, and blow them away... FOR REAL, yo. If you
want to move the modified color to the colors.xml in the package's
res dir, feel free... but hardcoding it in the layout = breaks themeing.

PS3:  Put back the lower case-ability of the date text, and cleaned whitespaces

Revert "Changed expanded clock to blue, and allow lower case date"

This reverts commit fa3c22f54008709a2862a379cc9f73263598c4e8.

Conflicts:

	packages/SystemUI/res/values/styles.xml

Change-Id: I7b334f121aa8736594e2a932ca5c1ba3ed4f31bb